* fixed queue cleanup (clean all queues for current kid) * fixed limit (to allow adding only max allowed kid count)

+++ src/main/java/com/zz/dats/kindergarten/controller/RestWeb.java
@@ -87,11 +87,22 @@ public class RestWeb {
     }
 
     @PostMapping("/kindergarten/dispatch/{id}")
-    public ModelAndView kindergartenAddKid(
+    public synchronized ModelAndView kindergartenAddKid(
         @PathVariable("id") Integer id
     ) {
-        kindergartenKidsRepository.addKidsFromQueue(id);
-        queueRepository.removeAddedKids(id);
+        Optional<KindergartenEntity> kindergartenOptional = kindergartenRepository.findById(id);
+
+        if (kindergartenOptional.isPresent()) {
+            KindergartenEntity kindergartenEntity = kindergartenOptional.get();
+            int limit = kindergartenEntity.getMaxKids() - kindergartenEntity.getKindergartenKidsById().size();
+
+            if (limit > 0) {
+                kindergartenKidsRepository.addKidsFromQueue(id, limit);
+                queueRepository.removeAddedKids(id);
+            }
+
+        }
+
 
         return new ModelAndView(String.format("redirect:/kindergarten/%d", id));
     }
